requirements.txt
fastapi
uvicorn
farm-haystack
transformers
torch
torchvision
faiss-cpu
elasticsearch
elasticsearch-dsl
pydantic
nltk
numpy
scikit-learn
scipy
sqlalchemy
httpx
pytest
pytest-asyncio
typing-extensions
datasets
sentence-transformers
tensorboard
python-multipart
pymupdf
blobfile
spacy
protobuf
sentencepiece
transformers[torch]
evaluate
typing
PyPDF2
pdf2image
pytesseract
